Add example for vm endpoint create-multiple
I didn't understand how to properly use that command. Same with other 3 developers that tried. I think is useful to add a small example showing that you can do it by separating each endpoint with commas.

I'm working on a UI that creates VM on the same cloud-service, so I need to automatize the endpoints creation and this is the best way to do it (in this case I'm running that commands with node inside a Meteor app).